refactor: go
